Added AIME Problems
Well I was gonnna do this later, but I was bored and hungry
because I was fasting lmao. Edited stuff to accomodate that after
adding the AIME folder from webscraping. Read the comments in
main.py if you would like to know more about how the naming
schemes work in the txts.